+++ src/main/java/ru/modgy/user/controller/UserController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.*;
 import ru.modgy.user.dto.NewUserDto;
 import ru.modgy.user.dto.UpdateUserDto;
 import ru.modgy.user.dto.UserDto;
+import ru.modgy.user.model.Roles;
 import ru.modgy.user.service.UserService;
 import ru.modgy.utility.UtilityService;
 
@@ -20,6 +21,7 @@ import java.util.Collection;
 @RequiredArgsConstructor
 public class UserController {
     private final UserService userService;
+    private final UtilityService utilityService;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
@@ -27,6 +29,7 @@ public class UserController {
                            @RequestBody @Valid NewUserDto userDto
 
     ) {
+        utilityService.checkHigherOrdinalRoleAccess(requesterId, userDto.getRole());
         log.info("UserController: POST/addUser, requesterId={}, user={}", requesterId, userDto);
         return userService.addUser(requesterId, userDto);
     }
@@ -35,6 +38,7 @@ public class UserController {
     public Collection<UserDto> getAllUsers(@RequestHeader(UtilityService.REQUESTER_ID_HEADER) Long requesterId,
                                            @RequestParam(value = "isActive", required = false) Boolean isActive
     ) {
+        utilityService.checkHigherOrEqualOrdinalRoleAccess(requesterId, Roles.ROLE_ADMIN);
         log.info("UserController: GET/getAllUsers, requesterId={}, isActive={}", requesterId, isActive);
         return userService.getAllUsers(requesterId, isActive);
     }
@@ -44,6 +48,9 @@ public class UserController {
     public UserDto getUserById(@RequestHeader(UtilityService.REQUESTER_ID_HEADER) Long requesterId,
                                @PathVariable(value = "id") long userId
     ) {
+        if (!utilityService.checkRequesterRequestsHimself(requesterId, userId)) {
+            utilityService.checkHigherOrEqualOrdinalRoleAccess(requesterId, userId);
+        }
         log.info("UserController: GET/getUserById, requesterId={}, userId={}", requesterId, userId);
         return userService.getUserById(requesterId, userId);
     }
@@ -53,6 +60,7 @@ public class UserController {
     public void deleteUserById(@RequestHeader(UtilityService.REQUESTER_ID_HEADER) Long requesterId,
                                @PathVariable("id") Long userId
     ) {
+        utilityService.checkHigherOrdinalRoleAccess(requesterId, userId);
         log.info("UserController: DELETE/deleteUserById, requesterId={}, userId={}", requesterId, userId);
         userService.deleteUserById(requesterId, userId);
     }
@@ -62,6 +70,9 @@ public class UserController {
                               @PathVariable(value = "id") Long userId,
                               @RequestBody @Valid UpdateUserDto updateUserDto
     ) {
+        if (!utilityService.checkRequesterRequestsHimself(requesterId, userId)) {
+            utilityService.checkHigherOrEqualOrdinalRoleAccess(requesterId, userId);
+        }
         log.info("UserController: PATCH/updateUser, requesterId={}, userId={}, requestBody={}",
                 requesterId, userId, updateUserDto);
         return userService.updateUser(requesterId, userId, updateUserDto);
@@ -72,6 +83,7 @@ public class UserController {
                                 @PathVariable(value = "id") Long userId,
                                 @RequestParam(value = "isActive", defaultValue = "true") Boolean isActive
     ) {
+        utilityService.checkHigherOrdinalRoleAccess(requesterId, userId);
         log.info("UserController: PATCH/setUserState, requesterId={}, userId={}, isActive={}",
                 requesterId, userId, isActive);
         return userService.setUserState(requesterId, userId, isActive);
